un-move the tests and update to fix some problems
