+++ tests/core/job/test_job_repository.py
@@ -19,8 +19,11 @@ from src.taipy.core.data._data_manager import _DataManager
 from src.taipy.core.data._data_manager_factory import _DataManagerFactory
 from src.taipy.core.data.csv import CSVDataNode
 from src.taipy.core.exceptions.exceptions import ModelNotFound
+from src.taipy.core.job._job_model import _JobModel
+from src.taipy.core.job._job_repository import _JobRepository
 from src.taipy.core.job._job_repository_factory import _JobRepositoryFactory
 from src.taipy.core.job.job import Job
+from src.taipy.core.job.status import Status
 from src.taipy.core.task._task_manager import _TaskManager
 from src.taipy.core.task._task_manager_factory import _TaskManagerFactory
 from src.taipy.core.task.task import Task
@@ -70,6 +73,18 @@ job = Job(JobId("id"), task, "submit_id", version="latest")
 job._subscribers = [f, A.f, A.g, A.h, A.B.f]
 job._exceptions = [traceback.TracebackException.from_exception(Exception())]
 
+job_model = _JobModel(
+    id=JobId("id"),
+    task_id=task.id,
+    status=Status(Status.SUBMITTED),
+    force=False,
+    submit_id="submit_id",
+    creation_date=job._creation_date.isoformat(),
+    subscribers=Job._serialize_subscribers(job._subscribers),
+    stacktrace=job._stacktrace,
+    version="latest",
+)
+
 
 class TestJobRepository:
     def test_save_and_load(self, tmpdir):
@@ -82,6 +97,15 @@ class TestJobRepository:
         _TaskManager._set(task)
         j = repository.load("id")
         assert j.id == job.id
+
+    def test_from_and_to_model(self):
+        repository = _JobRepositoryFactory._build_repository().repo
+        assert repository._to_model(job) == job_model
+        with pytest.raises(ModelNotFound):
+            repository._from_model(job_model)
+        _DataManager._set(data_node)
+        _TaskManager._set(task)
+        assert repository._from_model(job_model).id == job.id
 
     def test_save_and_load_with_sql_repo(self):
         Config.configure_global_app(repository_type="sql")
@@ -99,3 +123,19 @@ class TestJobRepository:
         _TaskManager._set(task)
         j = repository.load("id")
         assert j.id == job.id
+
+    def test_from_and_to_model_with_sql_repo(self):
+        Config.configure_global_app(repository_type="sql")
+
+        _DataManagerFactory._build_manager()._delete_all()
+        _TaskManagerFactory._build_manager()._delete_all()
+
+        repository = _JobRepositoryFactory._build_repository().repo._table
+        repository._delete_all()
+
+        assert repository._to_model(job) == job_model
+        with pytest.raises(ModelNotFound):
+            repository._from_model(job_model)
+        _DataManager._set(data_node)
+        _TaskManager._set(task)
+        assert repository._from_model(job_model).id == job.id
