+++ tests/core/task/test_task.py
@@ -9,22 +9,16 @@
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-import datetime
 from unittest import mock
 
 import pytest
 
-from src.taipy.core.common.alias import TaskId
 from src.taipy.core.config.data_node_config import DataNodeConfig
 from src.taipy.core.data._data_manager import _DataManager
-from src.taipy.core.data._data_manager_factory import _DataManagerFactory
 from src.taipy.core.data.csv import CSVDataNode
 from src.taipy.core.data.data_node import DataNode
 from src.taipy.core.data.in_memory import InMemoryDataNode
-from src.taipy.core.exceptions.exceptions import NonExistingDataNode
 from src.taipy.core.task._task_manager import _TaskManager
-from src.taipy.core.task._task_model import _TaskModel
-from src.taipy.core.task._task_repository_factory import _TaskRepositoryFactory
 from src.taipy.core.task.task import Task
 from taipy.config.common.scope import Scope
 from taipy.config.config import Config
@@ -49,74 +43,6 @@ def input():
 @pytest.fixture
 def input_config():
     return [DataNodeConfig("input_name_1"), DataNodeConfig("input_name_2"), DataNodeConfig("input_name_3")]
-
-
-data_node = CSVDataNode(
-    "test_data_node",
-    Scope.PIPELINE,
-    "dn_id",
-    "name",
-    "owner_id",
-    {"task_id"},
-    datetime.datetime(1985, 10, 14, 2, 30, 0),
-    [dict(timestamp=datetime.datetime(1985, 10, 14, 2, 30, 0), job_id="job_id")],
-    "latest",
-    None,
-    False,
-    {"path": "/path", "has_header": True},
-)
-
-task = Task(
-    "config_id",
-    {},
-    print,
-    [data_node],
-    [],
-    TaskId("id"),
-    owner_id="owner_id",
-    parent_ids={"parent_id"},
-    version="latest",
-)
-
-task_model = _TaskModel(
-    id="id",
-    owner_id="owner_id",
-    parent_ids=["parent_id"],
-    config_id="config_id",
-    input_ids=["dn_id"],
-    function_name=print.__name__,
-    function_module=print.__module__,
-    output_ids=[],
-    version="latest",
-    skippable=False,
-    properties={},
-)
-
-
-def test_from_and_to_model():
-    repository = _TaskRepositoryFactory._build_repository().repo  # type: ignore
-    assert repository._to_model(task) == task_model
-    with pytest.raises(NonExistingDataNode):
-        repository._from_model(task_model)
-    _DataManager._set(data_node)
-    t = repository._from_model(task_model)
-    assert isinstance(t, Task)
-    assert len(t.input) == 1
-
-
-def test_from_and_to_model_with_sql_repo():
-    Config.configure_global_app(repository_type="sql")
-
-    _DataManagerFactory._build_manager()._delete_all()
-    repository = _TaskRepositoryFactory._build_repository().repo._table  # type: ignore
-
-    assert repository._to_model(task) == task_model
-    with pytest.raises(NonExistingDataNode):
-        repository._from_model(task_model)
-    _DataManager._set(data_node)
-    t = repository._from_model(task_model)
-    assert isinstance(t, Task)
-    assert len(t.input) == 1
 
 
 def test_create_task():
